refactor: 장바구니에 관한 함수를 util 파일로 분리 및 적용
